trained chembl 600k modek, update to float16 ver

+++ models.py
@@ -31,7 +31,7 @@ class Encoder(nn.Module):
 
     def forward(self, x):
         # One-hot encode the input
-        x_one_hot = F.one_hot(x, self.vocab_size).float()
+        x_one_hot = F.one_hot(x.long(), self.vocab_size).to(torch.float16)
 
         # Apply dropout
         x_one_hot = self.dropout(x_one_hot)
@@ -83,7 +83,7 @@ class Decoder(nn.Module):
         batch_size = x.size(0)
 
         # One-hot encode the input
-        x_one_hot = F.one_hot(x, self.vocab_size).float()
+        x_one_hot = F.one_hot(x.long(), self.vocab_size).to(torch.float16)
 
         # Project latent to initial states
         states = F.relu(self.latent_to_states(latent))
